refactor: moved card view into cards package
